Safari-611.1.10/LayoutTests/editing/mac/spelling/advance-to-next-misspelling-expected.txt

Advancing to the next misspelling (cmd+; on macOS) should select the word 'mispelled'.
| <p>
|   id="start"
|   <br>
|   "<#selection-anchor>mispelled<#selection-focus>"
